child_name_app/requirements.txt
fastapi
uvicorn
openai
langchain
python-dotenv
pydantic
langchain-community
